refactor(skills): extract import pipeline into pkg/skills/importer
Pure relocation, no behavior change. Splits cmd/loom/skills_import.go
(1230 lines) into a transformational package + a thin CLI client so the
gRPC SkillsImportService (next step on this branch) can drive the same
pipeline server-side without import-cycling on cmd/loom/main.

Package layout (pkg/skills/importer):

  types.go      Skill, Reference, Frontmatter
  parse.go      ReadSkill, splitFrontmatter, readReferences,
                parseWhenToUseBullets, extractLinkedSkillNames,
                IsSafeSkillName, IsSkippedSkill
  keywords.go   BuildKeywords, BuildSlashCommands, MaxKeywords const,
                priority maps + ngram extraction
  render.go     RenderYAML, ChooseDomain, DeriveTitle, BuildInstructions,
                kvPair / mappingNode YAML helpers
  classify.go   Classify (stateless variant: skill + seed taxonomy ->
                LLM -> validated parent_index_path),
                ParseClassifyResponse, SuggestedTaxonomies
  pipeline.go   Run(Config) — orchestrates phases 1–5 with OnBanner +
                OnResult callbacks for progress streaming. Result
                struct surfaces classify path, ref counts, dropped
                refs, metadata field counts so callers can render
                detailed progress lines without re-running phase 4.

Tests moved into the package alongside their subjects:
  classify_test.go — ParseClassifyResponse (12 cases, unchanged)
  parse_test.go    — splitFrontmatter, extractLinkedSkillNames,
                     parseWhenToUseBullets, IsSafeSkillName,
                     IsSkippedSkill
  render_test.go   — RenderYAML round-trips through skills.LoadSkill,
                     normalizeCrossSkillLinks, ChooseDomain, DeriveTitle

cmd/loom/skills_import.go shrinks from 1230 to 311 lines and is now
purely a cobra command + per-skill output formatter that consumes the
streamed SkillResult records via OnResult. Output format is byte-for-
byte identical to the previous in-process implementation:

  ==> import: 23 source skills | mode=write | output=~/.loom/skills | classifier=...
          [skip] agent-skill-builder (meta-tooling, not convertible)
  [ 1/23] teradata-adaptive-optimizer    domain=teradata  trigger=AUTO  ...
  ...
  ==> summary: 23 converted, 1 skipped, 0 failed
  ==> classifications: teradata/admin:4, teradata/performance:4, ...

cmd/loom/skills_classify.go shrinks from 251 to 83 lines, retaining
only the env-driven LLM provider construction (buildClassifyLLM,
classifyProviderInfo) which is CLI-specific. The prompt builder,
validator, and Classify entry point all live in the package.

DiscoveryReport now carries MaxNameWidth so the CLI can size the name
column from real data rather than the prior 38-char heuristic; output
columns now align to the longest skill name in the run (was: padded to
width 38 regardless).

Verified:
  - go test -tags fts5 -race ./pkg/skills/importer/... ./cmd/loom/...
    green (TestRunSkillsImport_HappyPath / DryRun / RefusesOverwrite +
    all moved unit tests)
  - just check green (gofmt + golangci-lint + race tests + build +
    gosec 0 issues)
  - End-to-end smoke: ./bin/loom skills import ~/Projects/skills
    --dry-run produces identical 23-skill output
